pyproject.toml
[tool.poetry]
name = "base.django"
version = "0.1.0"
description = ""
authors = ["Trung Nguyen <nttrung143@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
celery = "^5.0.2"
django-filter = "^2.4.0"
djangorestframework = "^3.12.2"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
redis = "^3.5.3"
rest-pandas = "^1.1.0"
Babel = "^2.8.0"
Django = "^3.1.3"
Pillow = "^8.0.1"
python-dotenv = "^0.15.0"
graphene-django = "^2.13.0"
drf-yasg = {extras = ["validation"], version = "^1.20.0"}
djangorestframework-simplejwt = "^4.6.0"
Faker = "^5.0.1"
djangorestframework-camel-case = "^1.2.0"
coverage = "^5.5"
flake8 = "^3.9.2"
selenium = "^3.141.0"
tox = "^3.23.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
coverage = "^5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"